+++ src/components/ItaliaTheme/Header/HeaderCenter.jsx
@@ -21,8 +21,7 @@ import {
 
 const HeaderCenter = () => {
   const subsite = useSelector((state) => state.subsite?.data);
-
-  let logoSubsite = subsite?.subsite_logo && (
+  const logoSubsite = subsite?.subsite_logo && (
     <figure className="icon">
       <img
         src={flattenToAppURL(subsite.subsite_logo.scales?.mini?.download)}